add validation check for cpm, update return type for accuracy

+++ src/main/java/com/github/creme332/controller/GameOverScreenController.java
@@ -78,7 +78,7 @@ public class GameOverScreenController {
 
         // display accuracy
         gameOverScreen.setAccuracy(
-                calc.accuracy(
+                (long) calc.accuracy(
                         model.getTypeText().length(),
                         model.getTotalMistakes()));
     }
